requirements.txt
torch>=2.2.0
transformers>=4.41.0
accelerate>=0.31.0
faiss-cpu>=1.7.4
numpy>=1.26.0
scikit-learn>=1.3.0
datasets>=2.19.0
sentencepiece>=0.1.99
typer>=0.12.3
pydantic>=2.7.1
uvloop>=0.19.0
PyYAML>=6.0.1

